Audit fixes: Arduino include paths, PID reverse-windup, and hardening
Critical:
- Cross-directory includes in src/ and src/arduino/ headers were bare filenames
  (e.g. "IClock.h"), which fail under the Arduino IDE's recursive-library layout
  (only src/ is on the include path) -> NO sketch compiled. Now prefixed with
  core/ and hal/. Added 'make check-includes' guard that compiles
  '#include <RobotArm.h>' with only -Isrc so this cannot regress.

High:
- PIDController output-limit anti-windup folded P/D into the integrator, causing
  reverse windup (full-scale wrong-sign output at zero error after a P/D-term
  saturation). Replaced with integral-term clamping (bounds kI*integral to the
  output range); can never flip the integrator's sign. Regression tests added.

Lower severity / hardening:
- TrapezoidProfile: std::sqrt/std::fabs -> sqrtf/fabsf (§4.3 float-only math)
- RobotArm and Joint are now non-copyable/non-movable (they hold self-referential
  internal pointers: per-joint slew limiters, owned PCA9685 outputs) -> deleted
  copy/move + static_asserts
- setMaxSpeed(<=0) now disables smoothing (Joint::useDirect) instead of silently
  keeping the old rate
- ServoCalibration offset doc corrected (offset is ADDED to the commanded angle)
- inverse2: clamp branches use the same epsilon band as the reachability test, so
  reachable and clamped are never both true; invariant test added

89 cases / 22695 assertions green; check-core-purity + check-includes green;
Arduino path still compiles+links+runs against stubs with only -Isrc.

+++ src/core/TrapezoidProfile.cpp
@@ -22,7 +22,7 @@ Shape shapeOf(const TrapezoidProfile::Constraints& c,
               const TrapezoidProfile::State& goal) {
     Shape s{};
     const float delta = goal.position - start.position;
-    s.distance = std::fabs(delta);
+    s.distance = fabsf(delta);
     s.dir = (delta > 0.0f) ? 1.0f : (delta < 0.0f ? -1.0f : 0.0f);
 
     // Degenerate / invalid constraints: treat as an instantaneous move.
@@ -50,7 +50,7 @@ Shape shapeOf(const TrapezoidProfile::Constraints& c,
     } else {
         // Too short to reach cruise speed: a triangular profile peaking at
         // sqrt(a*D), reached at exactly the halfway distance.
-        s.vPeak = std::sqrt(a * s.distance);
+        s.vPeak = sqrtf(a * s.distance);
         s.tAccel = s.vPeak / a;
         s.dAccel = s.distance * 0.5f;
         s.tCruise = 0.0f;
